+++ app/src/main/cpp/CommonUtils.h
@@ -4,7 +4,7 @@
 #include <cstddef>
 
 void mat4f_load_ortho(float left, float right, float bottom, float top, float near, float far, float* mat4f);
-void mat4f_load_rotation_z(int rotation, float* mat4f);
+void mat4f_load_rotation_z(float rotation, float* mat4f);
 void mat4f_load_scale(float scaleX, float scaleY, float scaleZ, float* mat4f);
 float aspect_ratio_correction(bool fillScreen, size_t backingWidth, size_t backingHeight, size_t width, size_t height);
